% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_crsfn}
\alias{da_crsfn}
\title{da_crsfn}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 433 rows and 10 columns.
}
\usage{
da_crsfn
}
\description{
da_crsfn
}
\keyword{datasets}
