% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_pedidos_revisao}
\alias{da_pedidos_revisao}
\title{da_pedidos_revisao}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 62 rows and 10 columns.
}
\usage{
da_pedidos_revisao
}
\description{
da_pedidos_revisao
}
\keyword{datasets}
