% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{dados_movs}
\alias{dados_movs}
\title{dados_movs}
\format{
An object of class \code{data.frame} with 27213 rows and 4 columns.
}
\usage{
dados_movs
}
\description{
dados_movs
}
\keyword{datasets}
